+++ skeleton/Renderable.cpp
@@ -17,8 +17,8 @@ Renderable::Renderable(rapidjson::Document& d) : Renderable() {
 		rapidjson::Value& coordArray = featuresArray[i]["geometry"]["coordinates"];
 
 		for (rapidjson::SizeType j = 0; j < coordArray.Size(); j++) {
-			float lng = coordArray[j][0].GetDouble() * M_PI / 180.0;
-			float lat = coordArray[j][1].GetDouble() * M_PI / 180.0;
+			float lng = (float)( coordArray[j][0].GetDouble() * M_PI / 180.f );
+			float lat = (float)( coordArray[j][1].GetDouble() * M_PI / 180.f );
 
 			verts.push_back(glm::vec3(sin(lng)*cos(lat), sin(lat), cos(lng)*cos(lat)) * (float) RADIUS_EARTH_MODEL);
 			colours.push_back(glm::vec3(0.f, 0.f, 0.f));
